… test/ExtendedXmlSerializer.Tests.Performance/Benchmarks.cs