It looks good:

|                            Type |                            Method |     Mean |     Error |   StdDev |
|-------------------------------- |---------------------------------- |---------:|----------:|---------:|
|       ExtendedXmlSerializerTest |   SerializationClassWithPrimitive | 38.78 us | 21.685 us | 1.189 us |
| LegacyExtendedXmlSerializerTest |   SerializationClassWithPrimitive | 59.94 us | 72.373 us | 3.967 us |
|               XmlSerializerTest |   SerializationClassWithPrimitive | 30.90 us |  8.589 us | 0.471 us |
|       ExtendedXmlSerializerTest | DeserializationClassWithPrimitive | 86.00 us | 14.669 us | 0.804 us |
| LegacyExtendedXmlSerializerTest | DeserializationClassWithPrimitive | 97.92 us | 12.978 us | 0.711 us |
|               XmlSerializerTest | DeserializationClassWithPrimitive | 67.30 us |  6.271 us | 0.344 us |